+++ src/main/scala/com/programmaticallyspeaking/ncd/nashorn/ClassScanner.scala
@@ -5,7 +5,7 @@ import com.programmaticallyspeaking.ncd.messaging.{Observer, SerializedSubject}
 import com.programmaticallyspeaking.ncd.nashorn.NashornDebuggerHost.{ScriptClassNamePrefix, wantedTypes}
 import com.sun.jdi.event.ClassPrepareEvent
 import com.sun.jdi.request.EventRequest
-import com.sun.jdi.{ClassType, ReferenceType, ThreadReference, VirtualMachine}
+import com.sun.jdi._
 import org.slf4s.Logging
 
 import scala.collection.concurrent.TrieMap
@@ -61,7 +61,12 @@ class ClassScanner(virtualMachine: XVirtualMachine, scripts: Scripts, scriptFact
     // scanner decide if scripts were added.
     val classTracker = new ClassTracker(virtualMachine.inner)
 
-    try f finally {
+    var thrown: Throwable = null
+    try f catch {
+      case t: Throwable =>
+        thrown = t
+        throw t
+    } finally {
       val addedClasses = classTracker.addedClasses()
       try {
         log.debug(s"Scanning ${addedClasses.size} classes to detect scripts added during code evaluation.")
@@ -69,6 +74,8 @@ class ClassScanner(virtualMachine: XVirtualMachine, scripts: Scripts, scriptFact
       } catch {
         case NonFatal(t) =>
           log.error("Class scanning after code evaluation failed", t)
+          if (thrown != null) thrown.addSuppressed(t)
+          else throw t
       }
     }
   }
@@ -104,11 +111,11 @@ class ClassScanner(virtualMachine: XVirtualMachine, scripts: Scripts, scriptFact
           log.warn(s"Found the $className type but it's a ${other.getClass.getName} rather than a ClassType")
       }
     } else {
-      def callback(maybeIdentifiedScript: Option[IdentifiedScript]) = maybeIdentifiedScript match {
+      def callback(maybeIdentifiedScript: Option[IdentifiedScript], lineLocations: Seq[Location]) = maybeIdentifiedScript match {
         case Some(identifiedScript) =>
           try {
             val script = scripts.suggest(identifiedScript.script)
-            val bls = breakableLocations.add(script, refType.allLineLocations().asScala)
+            val bls = breakableLocations.add(script, lineLocations)
             activeBreakpoints.addBreakableLocations(script, bls)
             scriptPublisher.publish(script)
           } catch {
@@ -184,5 +191,7 @@ class ClassScanner(virtualMachine: XVirtualMachine, scripts: Scripts, scriptFact
     wantedTypes.contains(name) || name.startsWith(ScriptClassNamePrefix)
   }
 
-  def typeByName(name: String) = foundWantedTypes.get(name)
+  def typeByName(name: String): Option[ClassType] = foundWantedTypes.get(name).orElse {
+    virtualMachine.classByName(name).collect { case ct: ClassType => ct }
+  }
 }