+++ src/main/scala/com/programmaticallyspeaking/ncd/nashorn/JDIExtensions.scala
@@ -1,7 +1,7 @@
 package com.programmaticallyspeaking.ncd.nashorn
 
 import com.programmaticallyspeaking.ncd.host.ScopeType
-import com.programmaticallyspeaking.ncd.infra.ScriptURL
+import com.programmaticallyspeaking.ncd.infra.{CompiledScript, ScriptURL}
 import com.programmaticallyspeaking.ncd.nashorn.NashornDebuggerHost.{EventHandler, EventHandlerKey}
 import com.sun.jdi._
 import com.sun.jdi.event.Event
@@ -63,21 +63,30 @@ object JDIExtensions {
       // surprising), where it is stated that the Java stratum doesn't use source paths and a path therefore is a
       // package-qualified file name in path form, whereas name is the unqualified file name (e.g.:
       // java\lang\Thread.java vs Thread.java).
-      val path = location.sourceName()
-      if (path == "<eval>") {
-        // For evaluated scripts, convert the type name into something that resembles a file URI.
-        val typeName = location.declaringType().name()
-        "eval:/" + typeName
-          .replace("jdk.nashorn.internal.scripts.", "")
-          .replace('.', '/')
-          .replace('\\', '/')
-          .replaceAll("[$^_]", "")
-          .replaceFirst("/eval/?$", "")
-      } else {
-        path // keep it simple
-      }
+      val sourceName = location.sourceName()
+      sourceNameToUrl(location.declaringType(), sourceName)
     }
 
+  }
+
+  private def sourceNameToUrl(refType: ReferenceType, sourceName: String): String = sourceName match {
+    case "<eval>" =>
+      // For evaluated scripts, convert the type name into something that resembles a file URI.
+      val typeName = refType.name()
+      "eval:/" + typeNameToUrl(typeName)
+    case CompiledScript(cs) =>
+      cs.url
+    case _ =>
+      sourceName // keep it simple
+  }
+
+  private def typeNameToUrl(typeName: String): String = {
+    typeName
+      .replace("jdk.nashorn.internal.scripts.", "")
+      .replace('.', '/')
+      .replace('\\', '/')
+      .replaceAll("[$^_]", "")
+      .replaceFirst("/eval/?$", "")
   }
 
   class ExtValue(v: Value) {
@@ -99,6 +108,12 @@ object JDIExtensions {
       // Generated script classes has a field named 'source'
       Option(refType.fieldByName("source"))
         .getOrElse(throw new Exception("Found no 'source' field in " + refType.name()))
+    }
+
+    def scriptURL: ScriptURL = ScriptURL.create(scriptPath)
+    private lazy val scriptPath: String = {
+      val sourceName = referenceType.sourceName()
+      sourceNameToUrl(referenceType, sourceName)
     }
 
     /**